requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
python-dateutil>=2.8.0
requests>=2.31.0